+++ packages/ecs-winston-format/test/uncaught-exception.js
@@ -17,28 +17,28 @@
 
 'use strict'
 
-const winston = require('winston')
 const ecsFormat = require('../') // @elastic/ecs-winston-format
+const winston = require('winston')
 
-const logger = winston.createLogger({
+// eslint-disable-next-line
+const log = winston.createLogger({
   level: 'info',
   format: ecsFormat(),
-  // Compare to:
-  // format: winston.format.combine(
-  //   winston.format.errors({stack: true, cause: true}),
-  //   winston.format.json()
-  // ),
   transports: [
     new winston.transports.Console({
-      handleExceptions: true,
-      handleRejections: true
+      handleExceptions: true
     })
   ]
 })
 
-logger.info('hi')
-logger.warn('look out', { foo: 'bar' })
+function funcb () {
+  const e = new Error('funcb boom')
+  e.code = 42
+  throw e
+}
 
-const err = new Error('boom', { cause: new Error('the cause') })
-err.code = 42
-logger.error('here is an exception', err)
+function funca () {
+  funcb()
+}
+
+funca()